+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased][unreleased]
 
+- Capture relevant error stack trace
+- Add catch function to query execution
+- Make query execution properly then-able
+
 ## [2.0.2][] - 2021-10-14
 
 - Support `null` as value for `update`